+++ Classes/DependencyInjection/Widgets/ConversionsPerMonthRegistration.php
@@ -1,0 +1,98 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "matomo_widgets" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\MatomoWidgets\DependencyInjection\Widgets;
+
+use Brotkrueml\MatomoWidgets\Extension;
+use Brotkrueml\MatomoWidgets\Widgets\Provider\GenericBarChartDataProvider;
+use Symfony\Component\DependencyInjection\Reference;
+use TYPO3\CMS\Backend\View\BackendViewFactory;
+use TYPO3\CMS\Dashboard\Widgets\BarChartWidget;
+
+/**
+ * @internal
+ */
+final class ConversionsPerMonthRegistration extends AbstractRegistration
+{
+    private const METHOD = 'Goals.get';
+    private const PARAMETERS_PARAMETERS = 'matomo_widgets.conversionsPerMonth.parameters';
+
+    protected $serviceIdSuffix = 'goals.conversionsPerMonth';
+
+    public function register(): void
+    {
+        if (! $this->matomoConfiguration->isWidgetActive('conversionsPerMonth')) {
+            return;
+        }
+
+        $this->defineParameters();
+        $this->registerDataProvider();
+        $this->registerWidget();
+    }
+
+    private function defineParameters(): void
+    {
+        $this->parameters->set(
+            self::PARAMETERS_PARAMETERS,
+            [
+                'period' => 'month',
+                'date' => 'last12',
+                'showColumns' => 'nb_conversions',
+            ],
+        );
+    }
+
+    private function registerDataProvider(): void
+    {
+        $this->services
+            ->set($this->buildServiceDataProviderId(), GenericBarChartDataProvider::class)
+            ->arg('$connectionConfiguration', new Reference($this->connectionConfigurationId))
+            ->arg('$method', self::METHOD)
+            ->arg(
+                '$barLabel',
+                Extension::LANGUAGE_PATH_DASHBOARD . ':conversions',
+            )
+            ->arg('$backgroundColour', '#69bbb5')
+            ->arg('$parameters', '%' . self::PARAMETERS_PARAMETERS . '%');
+    }
+
+    private function registerWidget(): void
+    {
+        $localisedTitle = Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.goals.conversionsPerMonth.title';
+        $title = $this->matomoConfiguration->siteTitle !== ''
+            ? \sprintf('%s: %s', $this->matomoConfiguration->siteTitle, 'Conversions per month')
+            : $localisedTitle;
+
+        $this->services
+            ->set($this->buildServiceWidgetId(), BarChartWidget::class)
+            ->arg('$dataProvider', new Reference($this->buildServiceDataProviderId()))
+            ->arg('$backendViewFactory', new Reference(BackendViewFactory::class))
+            ->arg(
+                '$options',
+                [
+                    'siteTitle' => $this->matomoConfiguration->siteTitle,
+                    'title' => $localisedTitle,
+                ],
+            )
+            ->tag(
+                'dashboard.widget',
+                [
+                    'identifier' => $this->buildWidgetIdentifier(),
+                    'groupNames' => 'matomo',
+                    'title' => $title,
+                    'description' => Extension::LANGUAGE_PATH_DASHBOARD . ':widgets.goals.conversionsPerMonth.description',
+                    'iconIdentifier' => self::ICON_IDENTIFIER,
+                    'height' => 'medium',
+                    'width' => 'small',
+                ],
+            );
+    }
+}