+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Widget "Conversions per month" (#52)
+
 ## [3.2.3] - 2025-09-20
 
 ### Fixed